# Treat positive and negative 0 as equivalent when hashing