--- a/tests/testthat/test-hash.R
+++ b/tests/testthat/test-hash.R
@@ -135,7 +135,7 @@ test_that("can hash complex vectors", {
   )
 })
 
-test_that("hash treats positive and negative 0 as equivalent", {
+test_that("hash treats positive and negative 0 as equivalent (#637)", {
   expect_equal(vec_hash(-0), vec_hash(0))
 })