Closes #637 

We now treat positive and negative 0 as equivalent, similar to how R does
https://github.com/wch/r-source/blob/800fa450bca0bef95d53ce705cbc70f6075fdec0/src/main/unique.c#L115

``` r
vctrs::vec_unique(c(0, -0))
#> [1] 0
```

The way that R does, by creating a tmp variable with `double tmp = (xi == 0.0) ? 0.0 : xi;` is slightly slower than what I do by skipping the re-assignment when the value is not 0.

FYI, integers do not have this problem. `0` and `-0` look the same at the C level.

This has minimal performance implications (it looks like it got faster but I think that's just proof that it doesn't have much of an impact, and you are just seeing noise).

``` r
library(vctrs)
x <- rep(1, times = 1e7)

vec_hash <- vctrs:::vec_hash
```

Before:

``` r
bench::mark(vec_hash(x), iterations = 100)
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_hash(x)    117ms    131ms      7.28    38.1MB     7.58
```

After:

``` r
bench::mark(vec_hash(x), iterations = 100)
#> # A tibble: 1 x 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_hash(x)    119ms    124ms      8.01    38.1MB     8.34
```