@DavisVaughan 2019-10-25
🤔 I'm not quite sure what's happening with the azure linux builds. 

It says rlang 0.4.1.9000 is installed when remotes and rcmdcheck are installed

https://dev.azure.com/r-lib/vctrs/_build/results?buildId=601&view=logs&jobId=bb1c2637-64c6-57bd-9ea6-93823b2df951&taskId=4a1a54a7-28b3-5508-d0b8-dcd62167bb50&lineStart=254&lineEnd=255&colStart=1&colEnd=1

But then when the vctrs deps are actually installed it pulls in the 0.4.0 binary from rstudio package manager

https://dev.azure.com/r-lib/vctrs/_build/results?buildId=601&view=logs&jobId=bb1c2637-64c6-57bd-9ea6-93823b2df951&taskId=4a1a54a7-28b3-5508-d0b8-dcd62167bb50&lineStart=387&lineEnd=388&colStart=1&colEnd=1

So then later in the check it says rlang 0.4.0 is loaded, but rlang >= 0.4.1 is required

https://dev.azure.com/r-lib/vctrs/_build/results?buildId=601&view=logs&jobId=bb1c2637-64c6-57bd-9ea6-93823b2df951&taskId=9c62e7ff-391e-517d-ede1-af63bc6a25da&lineStart=62&lineEnd=63&colStart=1&colEnd=1

This must be some weird case where RPM hasn't refreshed to pull in the latest new CRAN rlang, but for some reason it isn't triggering a source install